workspace/requirements.txt
Django>=4.2
python-nmap>=0.7.1
requests>=2.31.0
celery>=5.3.0
redis>=5.0.0
django-celery-results>=2.5.0
django-celery-beat>=2.5.0
scapy>=2.5.0
psycopg2-binary>=2.9.9  # Optional if using PostgreSQL
